requirements.txt
asgiref==3.8.1
Django==5.0.7
django-etc==1.4.0
django-hitcount==1.3.5
django-modeltranslation==0.19.9
pillow==10.4.0
python-gettext==5.0
sqlparse==0.5.1
typing_extensions==4.12.2
